Refactor: Restructure into modular MCP architecture
Major restructuring following Model Context Protocol best practices:

## Changes

### Architecture
- Split monolithic 2,991-line main.py into 20 modular files
- Reduced main.py to 44 lines (server initialization only)
- Organized code by domain: subjects, characters, persons, users, collections, revisions, indices
- Implemented clean registration pattern for FastMCP components

### Bug Fixes
- Fixed EpisodeCollectionType enum (was 0=None,1=Watched; now 1=Wish,2=Done,3=Dropped)
- Fixed BANGUMI_TOKEN authentication (changed from static import to dynamic os.getenv)
- Fixed TextContent API usage (changed mimeType parameter to type="text")

### New Structure
- src/config.py - Configuration constants
- src/enums.py - 8 API enum definitions
- src/utils/api_client.py - HTTP client & error handling
- src/utils/formatters.py - Data formatting utilities
- src/resources/openapi_resource.py - OpenAPI spec resource
- src/tools/ - 55 tools split into 7 domain-specific modules
- src/prompts/workflow_prompts.py - 3 composite prompts

### Documentation
- Rewrote README with bilingual (EN/CN) comprehensive documentation
- Added project architecture explanation
- Added development guide and testing instructions

### Testing
- Verified all 55 tools, 1 resource, 3 prompts register correctly
- Tested authentication and episode collection update functionality
- All components working as expected
